modularise
this should be good I think
