+++ src/dialog.js
@@ -1,0 +1,4 @@
+const dialogPolyfill = require('dialog-polyfill');
+
+var dialog = document.querySelector('dialog');
+dialogPolyfill.registerDialog(dialog);
